Reorganizing Listener/PeerConnection relations
The Listener class is actually a factory for PeerConnections and the same
properties valid for connection classes is also valid for listeners
as template for creating PeerConnection instances. This makes the
PeerFactory redundant and is removed. PeerConnection class is moved to
connection module and the peer module has been removed.