Make the sockets have a freely usable "data" property
All sockets can now have a data object associated with them. This property
can be used freely and exist purely to be able to associate a connection
with an object which can be accessed through the eventhandlers. This also
removes the need for a special listener property in peerconnections, and
both this property and the overriding close() method has been removed.